+++ db/hc6/app/HC6.hcapp_syncUserData.StoredProcedure.sql
@@ -325,6 +325,7 @@ BEGIN
         k.KennelPaymentUrlExpires3                                          AS kennelPaymentUrlExpires3,
         k.KennelPaymentMemberSurcharge3                                     AS kennelPaymentMemberSurcharge3,
         k.KennelPaymentNonMemberSurcharge3                                  AS kennelPaymentNonMemberSurcharge3,
+        k.TrailSymbolsConfigJson                                            AS trailSymbolsConfigJson,
         k.removed                                                           AS removed,
         CONVERT(NVARCHAR(50), CAST(k.updatedAt AS DATETIME2))               AS updatedAt
     FROM HC.Kennel k