+++ mobile-app/lib/data/hc3_services/kennels/kennels_service.dart
@@ -77,6 +77,7 @@ class KennelsTableHelper extends BaseTableHelper<AppDomainType>
   final String colRunCountStartDate = 'runCountStartDate';
   final String colKennelMismanagementTeam = 'kennelMismanagementTeam';
   final String colDistancePreference = 'distancePreference';
+  final String colTrailSymbolsConfigJson = 'trailSymbolsConfigJson';
 
   @override
   Future<dynamic> createTable(
@@ -143,6 +144,7 @@ class KennelsTableHelper extends BaseTableHelper<AppDomainType>
             $colRunCountStartDate TEXT,
             $colKennelMismanagementTeam TEXT,
             $colDistancePreference INT,
+            $colTrailSymbolsConfigJson TEXT,
             $colUpdatedAt TEXT NOT NULL,
             $colRemoved INT NOT NULL,
             $colUpdatedAtValue INT NULL