+++ portal/lib/admin_pages/kennel_page_new/kennel_page_new_ui.dart
@@ -24,6 +24,7 @@ part 'pages/kennel_hash_cash_page/layout.dart';
 part 'pages/kennel_songs_page/layout.dart';
 part 'pages/kennel_logo_page/layout.dart';
 part 'pages/kennel_platform_admin_page/layout.dart';
+part 'pages/kennel_trail_symbols_page/layout.dart';
 part 'pages/kennel_page_new_widgets.dart';
 
 // ---------------------------------------------------------------------------
@@ -296,6 +297,8 @@ class _KennelTabBarView extends StatelessWidget {
         return KennelLogoTabContent(controller: controller);
       case KennelTabType.platformAdmin:
         return KennelPlatformAdminTabContent(controller: controller);
+      case KennelTabType.trailSymbols:
+        return KennelTrailSymbolsTabContent(controller: controller);
     }
   }
 }